Moved TypeError tests per PR discussion
